Policyを使って投稿の更新削除の制限を設定　するためcontrollerなどの修正

+++ app/Http/Controllers/PostController.php
@@ -69,8 +69,12 @@ class PostController extends Controller
   /**
    * Show the form for editing the specified resource.
    */
-  public function edit(Post $post)
+  public function edit(Request $request, Post $post)
   {
+    if ($request->user()->cannot('update', $post)) {
+      abort(403);
+    }
+
     return view('post.edit', compact('post'));
   }
 
@@ -79,6 +83,10 @@ class PostController extends Controller
    */
   public function update(Request $request, Post $post)
   {
+    if ($request->user()->cannot('update', $post)) {
+      abort(403);
+    }
+
     $inputs = $request->validate([
       'title' => 'required|max:255',
       'body' => 'required|max:1000',
@@ -103,8 +111,12 @@ class PostController extends Controller
   /**
    * Remove the specified resource from storage.
    */
-  public function destroy(Post $post)
+  public function destroy(Request $request, Post $post)
   {
+    if ($request->user()->cannot('delete', $post)) {
+      abort(403);
+    }
+
     $post->comments()->delete(); //投稿削除の時にコメントも削除
     $post->delete();
     return redirect()->route('post.index')->with('message', '投稿を削除しました');
